% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachatt.R
\name{attachatt}
\alias{attachatt}
\title{Attach attributes to network}
\usage{
attachatt(g, attdf, index, attvars)
}
\arguments{
\item{g}{igraph graph with named nodes (includes an attribute for 'name')}

\item{attdf}{Data frame containing desired attributes}

\item{index}{Character string; name of column in 'attdf' that contains the node names, to be used for matching to graph names}

\item{attvars}{Optional vector of variable name(s) in 'attdf' for attributes to be attached. If argument is missing, all variables in 'attdf' aside from the index column will be attached (with a warning message issued). To suppress warning, use attvars=NULL.}
}
\value{
The original network graph with attributes added
}
\description{
Attach one or more node attributes to a network object.
}
