% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attfile}
\alias{attfile}
\title{Attribute file}
\format{
A data frame with 328 rows and 4 variables:
\describe{
\item{PPID}{Participant ID number}
\item{gender}{F or M}
\item{race}{Black, EAsian, Latinx, SAsian, or White}
\item{age}{Age in years from 18-22}
\item{polorient}{Political orientation from 1 (most liberal) - 100 (most conservative)}
}
}
\usage{
attfile
}
\description{
Sample wide attribute file to accompany biology student network.
All data in this file was randomly generated (does not reflect
true attribute data of the biology students).
}
\keyword{datasets}
