% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmc.R
\name{gmc}
\alias{gmc}
\title{Grand mean center a vector}
\usage{
gmc(x)
}
\arguments{
\item{x}{Vector to be grand mean centered}
}
\value{
Vector of grand mean centered values
}
\description{
Centers each value of a vector on the mean of that vector.
}
\details{
For networks, useful for control variables (e.g., network size) in a model so that intercept represents average individual.
}
