% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeperm.R
\name{nodeperm}
\alias{nodeperm}
\title{Permute nodes}
\usage{
nodeperm(mat, mode = "directed", weighted = TRUE, diag = FALSE)
}
\arguments{
\item{mat}{Network matrix}

\item{mode}{Character string; How igraph should interpret matrix; default is 'directed' -- other common choice is 'undirected'}

\item{weighted}{Logical; if TRUE (default), network is weighted; if FALSE, network is not weighted}

\item{diag}{Logical; if FALSE (default), self-nominations are removed; if TRUE, network will include self-nominations}
}
\value{
An igraph object with the same underlying structure as original network, but randomized node labels
}
\description{
Permute node labels of a network
}
