% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pperm.R
\name{pperm}
\alias{pperm}
\title{Calculate p-perm}
\usage{
pperm(obs, permcoefs, plot = TRUE, coefname = "")
}
\arguments{
\item{obs}{Numeric; observed coefficient}

\item{permcoefs}{Numeric vector; coefficients derived from permuted networks}

\item{plot}{Logical; if TRUE (default), distribution plot will be included in output; if FALSE, only p-perm will be calculated}

\item{coefname}{Optional character string; used to specify type of coefficent for x-axis label of plot (if not provided, x-axis label will just be 'Coefficient')}
}
\value{
Either a single numeric value representing p-perm if plot=FALSE; otherwise a list including both the plot and p-perm value
}
\description{
Calculate two-sided exact p-perm value for a node permutation test, with optional plot of observed coefficient vs. distribution of coefficients from permuted networks
}
