% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale01.R
\name{rescale01}
\alias{rescale01}
\title{Rescale vector}
\usage{
rescale01(x, min = NULL, max = NULL)
}
\arguments{
\item{x}{Vector to be rescaled}

\item{min}{Optional numeric; minimum value possible that should be rescaled to 0 (otherwise will be scaled using minimum value represented in vector as 0)}

\item{max}{Optional numeric; maximum value possible that should be rescaled to 1 (otherwise will be scaled using maximum value represented in vector as 1)}
}
\value{
Vector rescaled to 0-1
}
\description{
Rescales a vector to 0-1.
}
\details{
For networks, useful if weights/measures of tie strengths are on different scales.
}
