% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tierecip.R
\name{tierecip}
\alias{tierecip}
\title{Calculate node-level tie reciprocity}
\usage{
tierecip(el, type = "all")
}
\arguments{
\item{el}{Edgelist with "from" and "to" columns}

\item{type}{Character string; either "all" (default; calculates node-level reciprocity using all nodes and ties in edgelist) or "complete" (calculates reciprocity only based on nominees who had the potential to reciprocate because they completed the network survey--this assumes that all individuals who completed the survey and no individuals who did not are listed in the 'from' column)}
}
\value{
A dataframe containing node IDs and tie reciprocity proportions
}
\description{
Calculate proportion of reciprocal ties out of total ties for each node from an edgelist.
}
\details{
Base calculations either on all nodes and ties in edgelist or only on those who had the opportunity to nominate ties (e.g., completed the network survey).
Self-loops (individuals nominating themselves) are not included.
Individuals who nominated 0 ties or just themselves have a reciprocity of NaN.
}
